Remove metions about local/remote fortune access in CRUD factory. Remote adapter moved to separate file
