+++ lib/crud-factory.js
@@ -2,20 +2,13 @@ var util = require("./util"),
     _ = require("lodash"),
     when = require("when");
 
-
-//TODO: break down into local/remote cruds
-
 module.exports = function(app, resources,actionNames){
   var crud = {
     fancy: {},
     simple: {}
   };
   
-  if(_.isObject(app)){
-    initWithLocalFortune();
-  }else{
-    initWithRemoteFortune();
-  }
+  init();
 
   //check if the query is for a whole, possibly filtered collection
   function isCollectionQuery(query){
@@ -31,7 +24,7 @@ module.exports = function(app, resources,actionNames){
     }
   }
 
-  function wrapDirectCall(method, resource, request, options){
+  function wrapCall(method, resource, request, options){
     options = options || {};
     return function(cb){
       var dispatch = function(){
@@ -98,21 +91,18 @@ module.exports = function(app, resources,actionNames){
   }
 
   function makeAction(name, resource){
-    /*NOTE: in all of these, 'options' should thought of as fortune-client settings rather than
-     * something that is meant to go through directly into fortune direct calls
-     */
     return {
       get: function(query,options){
-        return wrapDirectCall(name, resource.route, toDirectRequest(query,options),options);
+        return wrapCall(name, resource.route, toDirectRequest(query,options),options);
       },
       destroy: function(query,options){
-        return wrapDirectCall(name, resource.route, toDirectRequest(query,options),options);
+        return wrapCall(name, resource.route, toDirectRequest(query,options),options);
       },
       create: function(data, options){
         if(!(_.isPlainObject(data)||_.isArray(data))){
           throw new Error("First argument must be an object or an array.");
         }
-        return wrapDirectCall(name,
+        return wrapCall(name,
                               resource.route,
                               toDirectRequest(null, options, resource.route,data),
                               options);
@@ -120,14 +110,14 @@ module.exports = function(app, resources,actionNames){
       replace: function(id, data, options){
         if(isCollectionQuery(id)) throw new Error("Collection queries not allowed in replace!");
 
-        return wrapDirectCall(name,
+        return wrapCall(name,
                               resource.route,
                               toDirectRequest(id,options,resource.route, data),
                               options);
       },
       update: function(id, data, options){
         if(isCollectionQuery(id)) throw new Error("Collection queries not allowed in update!");
-        return wrapDirectCall(name, resource.route, toDirectRequest(id,options,null,data),options);
+        return wrapCall(name, resource.route, toDirectRequest(id,options,null,data),options);
       }
     }[name];
   }
@@ -157,7 +147,7 @@ module.exports = function(app, resources,actionNames){
     };
   }
 
-  function initWithLocalFortune(){
+  function init(){
     _.each(resources, function(resource){
       var plural = util.toCapitalisedCamelCase(resource.route),
           singular = util.toCapitalisedCamelCase(resource.name);
@@ -181,10 +171,6 @@ module.exports = function(app, resources,actionNames){
         return makeAction(name, {route: route}).apply(crud, _.rest(arguments));
       };
     });
-  }
-
-  function initWithRemoteFortune(){
-    throw new Error("initWithRemoteFortune not implemented");
   }
 
   return crud;
