+++ lib/fortune-client.js
@@ -3,7 +3,8 @@ var when = require("when"),
     request = require("request"),
     resourceLinker = require("./resource-linker"),
     crypto = require("crypto"),
-    _ = require("lodash")
+    _ = require("lodash"),
+    remoteAdapter = require("./remote-adapter");
 
 module.exports = function(apps){
   var self = {
@@ -37,73 +38,11 @@ module.exports = function(apps){
       }else if(_.isString(app)){
         return requesto.get(app + "/resources").then(function(data){
           resources = JSON.parse(data.body).resources;
-
-          remoteAdapter = {
-            'get': function(resource, req) {
-              var deferred = when.defer();
-              request({
-                method: "GET",
-                uri: [app, resource, (req.params || {}).id].join('/'),
-                qs: req.query
-              }, function(err, data) {
-                return err ? deferred.reject(err) : deferred.resolve({ body: JSON.parse(data.body) });
-              })
-              return deferred.promise;
-            },
-            'create': function(resource, req) {
-              var deferred = when.defer();
-              request({
-                method: "POST",
-                uri: [app, resource].join('/'),
-                json: true,
-                body: req.body
-              }, function(err, data) {
-                return err ? deferred.reject(err) : deferred.resolve({ body: data.body });
-              })
-              return deferred.promise;
-            },
-            'destroy': function(resource, req) {
-              var deferred = when.defer();
-
-              request({
-                method: "DELETE",
-                uri: [app, resource, (req.params || {}).id].join('/'),
-                json: true
-              }, function(err, data) {
-                return err ? deferred.reject(err) : deferred.resolve({ body: data.body });
-              })
-              return deferred.promise;
-            },
-            'replace': function(resource, req) {
-              var deferred = when.defer();
-              request({
-                method: "PUT",
-                uri: [app, resource, (req.params || {}).id].join('/'),
-                json: true,
-                body: req.body
-              }, function(err, data) {
-                return err ? deferred.reject(err) : deferred.resolve({ body: data.body });
-              })
-              return deferred.promise;
-            },
-            'update': function(resource, req) {
-              var deferred = when.defer();
-              request({
-                method: "PATCH",
-                uri: [app, resource, (req.params || {}).id].join('/'),
-                json: true,
-                body: req.body
-              }, function(err, data) {
-                return err ? deferred.reject(err) : deferred.resolve({ body: data.body });
-              })
-              return deferred.promise;
-            },
-          }
           
-          router.registerResources(remoteAdapter, resources);
+          router.registerResources(remoteAdapter(app), resources);
 
           return self.resources = _.union(self.resources, resources);
-        });
+        }).catch(function(err) { console.trace(err) });
       }else{
         throw new Error("App must be a function or an object, got " + typeof app);
       }
